better organize/isolate config/components
the life of a monorepo
