+++ config/rollup.config.js
@@ -1,12 +1,12 @@
 import typescript from '@rollup/plugin-typescript'
 
 export default {
-  input: 'src/index.ts',
+  input: 'src/lib/index.ts',
   output: {
     name: 'Outerbase',
     file: 'dist/umd/bundle.js',
     format: 'umd',
     sourcemap: true,
   },
-  plugins: [typescript()],
+  plugins: [typescript({ tsconfig: 'config/tsconfig.umd.json' })],
 }
