+++ config/tsconfig.cjs.json
@@ -1,7 +1,8 @@
 {
   "extends": "../tsconfig",
+  "include": ["../src/lib"],
   "compilerOptions": {
-    "module": "commonjs" /* Specify module code generation: 'none', 'commonjs', 'amd', 'system', 'umd', 'es2015', or 'ESNext'. */,
-    "outDir": "../dist/cjs" /* Redirect output structure to the directory. */
+    "outDir": "../dist/cjs",
+    "sourceMap": true
   }
 }
