+++ config/tsconfig.cli.json
@@ -1,14 +1,8 @@
 {
-  "include": ["../cli/**/*", "../src/**/*"],
+  "extends": "../tsconfig",
+  "include": ["../src/cli"],
   "compilerOptions": {
-    "rootDir": "..",
-    "outDir": "../dist",
-    "strict": true,
     "target": "es6",
-    "module": "commonjs",
-    "sourceMap": true,
-    "esModuleInterop": true,
-    "resolveJsonModule": true,
-    "moduleResolution": "node"
+    "outDir": "../dist/cli"
   }
 }
