+++ config/tsconfig.esm.json
@@ -1,7 +1,9 @@
 {
   "extends": "../tsconfig",
+  "include": ["../src/lib"],
   "compilerOptions": {
-    "module": "ESNext" /* Specify module code generation: 'none', 'commonjs', 'amd', 'system', 'umd', 'es2015', or 'ESNext'. */,
-    "outDir": "../dist/esm" /* Redirect output structure to the directory. */
+    "module": "ESNext",
+    "outDir": "../dist/esm",
+    "sourceMap": true
   }
 }
