+++ config/tsconfig.umd.json
@@ -1,7 +1,7 @@
 {
   "extends": "../tsconfig",
+  "include": ["../src/lib"],
   "compilerOptions": {
-    "module": "commonjs" /* Specify module code generation: 'none', 'commonjs', 'amd', 'system', 'umd', 'es2015', or 'ESNext'. */,
-    "declaration": false /* Generates corresponding '.d.ts' file. */
+    "module": "ESNext"
   }
 }
