+++ package.json
@@ -10,9 +10,9 @@
     "ezql": "dist/cli/index.js"
   },
   "scripts": {
-    "test:watch": "jest --watch",
-    "test": "jest",
-    "test:cov": "jest --coverage",
+    "test:watch": "jest --config config/jest.config.js --watch",
+    "test": "jest --config config/jest.config.js",
+    "test:cov": "jest --config config/jest.config.js --coverage",
     "watch": "npx tsc -w",
     "watch:cli": "npx tsc -w -p config/tsconfig.cli.json",
     "build": "rm -rf dist && concurrently npm:build:*",
@@ -20,7 +20,7 @@
     "build:cjs": "tsc -p config/tsconfig.cjs.json",
     "build:esm": "tsc -p config/tsconfig.esm.json",
     "build:types": "tsc -p config/tsconfig.types.json",
-    "build:umd": "rollup --config --bundleConfigAsCjs",
+    "build:umd": "rollup --config config/rollup.config.js --bundleConfigAsCjs",
     "update:demo": "npm run build:umd && node demo/copy-assets.mjs",
     "prepare": "husky install",
     "postinstall": "npm run build:cli"
